refactor: rename package servercache to cache
