+++ internal/cache/cache.go
@@ -1,4 +1,4 @@
-package servercache
+package cache
 
 import (
 	"context"
@@ -9,7 +9,6 @@ import (
 	"k8s.io/klog/v2"
 
 	"github.com/hetznercloud/hcloud-cloud-controller-manager/internal/metrics"
-	"github.com/hetznercloud/hcloud-go/v2/hcloud"
 )
 
 type Mode string
@@ -72,27 +71,6 @@ type Cache[T any] struct {
 	byName map[string]*entry[T]
 
 	mu sync.Mutex
-}
-
-func NewServerCache(client *hcloud.Client, defaultMode Mode, defaultTTL time.Duration) *Cache[hcloud.Server] {
-	return newCache[hcloud.Server](
-		func(ctx context.Context, id int64) (*hcloud.Server, error) {
-			value, _, err := client.Server.GetByID(ctx, id)
-			return value, err
-		},
-		func(ctx context.Context, name string) (*hcloud.Server, error) {
-			value, _, err := client.Server.GetByName(ctx, name)
-			return value, err
-		},
-		func(ctx context.Context) ([]*hcloud.Server, error) {
-			values, err := client.Server.All(ctx)
-			return values, err
-		},
-		func(value *hcloud.Server) int64 { return value.ID },
-		func(value *hcloud.Server) string { return value.Name },
-		defaultMode,
-		defaultTTL,
-	)
 }
 
 func newCache[T any](
