+++ internal/config/config_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"github.com/hetznercloud/hcloud-cloud-controller-manager/internal/servercache"
+	"github.com/hetznercloud/hcloud-cloud-controller-manager/internal/cache"
 	"github.com/hetznercloud/hcloud-cloud-controller-manager/internal/testsupport"
 	"github.com/hetznercloud/hcloud-go/v2/hcloud"
 )
@@ -27,7 +27,7 @@ func TestRead(t *testing.T) {
 				Robot:    RobotConfiguration{CacheTimeout: 5 * time.Minute},
 				Metrics:  MetricsConfiguration{Enabled: true, Address: ":8233"},
 				Instance: InstanceConfiguration{AddressFamily: AddressFamilyIPv4},
-				Cache:    CacheConfiguration{Mode: servercache.ModeAll, TTL: 10 * time.Second},
+				Cache:    CacheConfiguration{Mode: cache.ModeAll, TTL: 10 * time.Second},
 				Network: NetworkConfiguration{
 					AttachedCheckEnabled: true,
 				},
@@ -51,7 +51,7 @@ func TestRead(t *testing.T) {
 				Robot:        RobotConfiguration{CacheTimeout: 5 * time.Minute},
 				Metrics:      MetricsConfiguration{Enabled: true, Address: ":8233"},
 				Instance:     InstanceConfiguration{AddressFamily: AddressFamilyIPv4},
-				Cache:        CacheConfiguration{Mode: servercache.ModeAll, TTL: 10 * time.Second},
+				Cache:        CacheConfiguration{Mode: cache.ModeAll, TTL: 10 * time.Second},
 				Network: NetworkConfiguration{
 					NameOrID:             "foobar",
 					AttachedCheckEnabled: true,
@@ -89,7 +89,7 @@ func TestRead(t *testing.T) {
 				},
 				Metrics:  MetricsConfiguration{Enabled: true, Address: ":8233"},
 				Instance: InstanceConfiguration{AddressFamily: AddressFamilyIPv4},
-				Cache:    CacheConfiguration{Mode: servercache.ModeAll, TTL: 10 * time.Second},
+				Cache:    CacheConfiguration{Mode: cache.ModeAll, TTL: 10 * time.Second},
 				Network: NetworkConfiguration{
 					AttachedCheckEnabled: true,
 				},
@@ -146,7 +146,7 @@ failed to read ROBOT_PASSWORD_FILE: open /tmp/hetzner-password: no such file or 
 				Robot:    RobotConfiguration{CacheTimeout: 5 * time.Minute},
 				Metrics:  MetricsConfiguration{Enabled: true, Address: ":8233"},
 				Instance: InstanceConfiguration{AddressFamily: AddressFamilyIPv4},
-				Cache:    CacheConfiguration{Mode: servercache.ModeAll, TTL: 10 * time.Second},
+				Cache:    CacheConfiguration{Mode: cache.ModeAll, TTL: 10 * time.Second},
 				Network: NetworkConfiguration{
 					AttachedCheckEnabled: true,
 				},
@@ -176,7 +176,7 @@ failed to read ROBOT_PASSWORD_FILE: open /tmp/hetzner-password: no such file or 
 				Robot:    RobotConfiguration{CacheTimeout: 5 * time.Minute},
 				Metrics:  MetricsConfiguration{Enabled: false, Address: "127.0.0.1:9999"},
 				Instance: InstanceConfiguration{AddressFamily: AddressFamilyIPv4},
-				Cache:    CacheConfiguration{Mode: servercache.ModeAll, TTL: 10 * time.Second},
+				Cache:    CacheConfiguration{Mode: cache.ModeAll, TTL: 10 * time.Second},
 				Network: NetworkConfiguration{
 					AttachedCheckEnabled: true,
 				},
@@ -208,7 +208,7 @@ failed to read ROBOT_PASSWORD_FILE: open /tmp/hetzner-password: no such file or 
 				},
 				Metrics:  MetricsConfiguration{Enabled: true, Address: ":8233"},
 				Instance: InstanceConfiguration{AddressFamily: AddressFamilyIPv4},
-				Cache:    CacheConfiguration{Mode: servercache.ModeAll, TTL: 10 * time.Second},
+				Cache:    CacheConfiguration{Mode: cache.ModeAll, TTL: 10 * time.Second},
 				Network: NetworkConfiguration{
 					AttachedCheckEnabled: true,
 				},
@@ -241,7 +241,7 @@ failed to read ROBOT_PASSWORD_FILE: open /tmp/hetzner-password: no such file or 
 				},
 				Metrics:  MetricsConfiguration{Enabled: true, Address: ":8233"},
 				Instance: InstanceConfiguration{AddressFamily: AddressFamilyIPv4},
-				Cache:    CacheConfiguration{Mode: servercache.ModeAll, TTL: 10 * time.Second},
+				Cache:    CacheConfiguration{Mode: cache.ModeAll, TTL: 10 * time.Second},
 				Network: NetworkConfiguration{
 					AttachedCheckEnabled: true,
 				},
@@ -262,7 +262,7 @@ failed to read ROBOT_PASSWORD_FILE: open /tmp/hetzner-password: no such file or 
 				Robot:    RobotConfiguration{CacheTimeout: 5 * time.Minute},
 				Metrics:  MetricsConfiguration{Enabled: true, Address: ":8233"},
 				Instance: InstanceConfiguration{AddressFamily: AddressFamilyIPv6},
-				Cache:    CacheConfiguration{Mode: servercache.ModeAll, TTL: 10 * time.Second},
+				Cache:    CacheConfiguration{Mode: cache.ModeAll, TTL: 10 * time.Second},
 				Network: NetworkConfiguration{
 					AttachedCheckEnabled: true,
 				},
@@ -284,7 +284,7 @@ failed to read ROBOT_PASSWORD_FILE: open /tmp/hetzner-password: no such file or 
 				Robot:    RobotConfiguration{CacheTimeout: 5 * time.Minute},
 				Metrics:  MetricsConfiguration{Enabled: true, Address: ":8233"},
 				Instance: InstanceConfiguration{AddressFamily: AddressFamilyIPv4},
-				Cache:    CacheConfiguration{Mode: servercache.ModeAll, TTL: 10 * time.Second},
+				Cache:    CacheConfiguration{Mode: cache.ModeAll, TTL: 10 * time.Second},
 				LoadBalancer: LoadBalancerConfiguration{
 					Enabled:               true,
 					PrivateIngressEnabled: true,
@@ -308,7 +308,7 @@ failed to read ROBOT_PASSWORD_FILE: open /tmp/hetzner-password: no such file or 
 				Robot:    RobotConfiguration{CacheTimeout: 5 * time.Minute},
 				Metrics:  MetricsConfiguration{Enabled: true, Address: ":8233"},
 				Instance: InstanceConfiguration{AddressFamily: AddressFamilyIPv4},
-				Cache:    CacheConfiguration{Mode: servercache.ModeAll, TTL: 10 * time.Second},
+				Cache:    CacheConfiguration{Mode: cache.ModeAll, TTL: 10 * time.Second},
 				LoadBalancer: LoadBalancerConfiguration{
 					Enabled:               true,
 					PrivateIngressEnabled: true,
@@ -336,7 +336,7 @@ failed to read ROBOT_PASSWORD_FILE: open /tmp/hetzner-password: no such file or 
 				Robot:    RobotConfiguration{CacheTimeout: 5 * time.Minute},
 				Metrics:  MetricsConfiguration{Enabled: true, Address: ":8233"},
 				Instance: InstanceConfiguration{AddressFamily: AddressFamilyIPv4},
-				Cache:    CacheConfiguration{Mode: servercache.ModeAll, TTL: 10 * time.Second},
+				Cache:    CacheConfiguration{Mode: cache.ModeAll, TTL: 10 * time.Second},
 				Network: NetworkConfiguration{
 					AttachedCheckEnabled: true,
 				},
@@ -467,7 +467,7 @@ func TestHCCMConfiguration_Validate(t *testing.T) {
 			fields: fields{
 				HCloudClient: HCloudClientConfiguration{Token: "jr5g7ZHpPptyhJzZyHw2Pqu4g9gTqDvEceYpngPf79jN_NOT_VALID_dzhepnahq"},
 				Instance:     InstanceConfiguration{AddressFamily: AddressFamilyIPv4},
-				Cache:        CacheConfiguration{Mode: servercache.Mode("foobar")},
+				Cache:        CacheConfiguration{Mode: cache.Mode("foobar")},
 			},
 			wantErr: errors.New("invalid value for \"HCLOUD_SERVER_CACHE_MODE\", expect one of: all,one,off"),
 		},
@@ -563,11 +563,11 @@ func TestHCCMConfiguration_Validate(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			cache := tt.fields.Cache
-			if cache.Mode == "" {
+			cacheConfig := tt.fields.Cache
+			if cacheConfig.Mode == "" {
 				// Mirror the default set by Read() so cases that don't
 				// exercise cache validation stay valid.
-				cache.Mode = servercache.ModeAll
+				cacheConfig.Mode = cache.ModeAll
 			}
 			c := HCCMConfiguration{
 				HCloudClient: tt.fields.HCloudClient,
@@ -577,7 +577,7 @@ func TestHCCMConfiguration_Validate(t *testing.T) {
 				LoadBalancer: tt.fields.LoadBalancer,
 				Network:      tt.fields.Network,
 				Route:        tt.fields.Route,
-				Cache:        cache,
+				Cache:        cacheConfig,
 			}
 			err := c.Validate()
 			if tt.wantErr == nil {
